pyproject.toml
[tool.poetry]
name = "livraria"
version = "0.1.0"
description = ""
authors = ["Marco André Mendes <marcoandre@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
